ragchatbot/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
llama-index-core>=0.12,<0.13
llama-index-llms-openrouter>=0.3,<0.5
llama-index-embeddings-huggingface>=0.5,<0.7
llama-index-readers-file>=0.4,<0.5
pydantic-settings>=2.0,<3.0
sse-starlette==2.2.1
